Feature/sumin#admin member#001

+++ src/main/java/com/tech/petfriends/login/controller/LoginController.java
@@ -20,6 +20,9 @@ import com.tech.petfriends.login.mapper.MemberMapper;
 import com.tech.petfriends.login.util.PasswordEncryptionService;
 import com.tech.petfriends.member.service.MemberService;
 import com.tech.petfriends.mypage.dto.GradeDto;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
 
 @Controller
 @RequestMapping("/login")
@@ -32,7 +35,9 @@ public class LoginController {
     private MemberService memberService;
    
     @GetMapping("/loginPage")
-    public String LoginPage(HttpServletRequest request) {
+    public String LoginPage(HttpServletRequest request, Model model) {
+    	String previousUrl = request.getHeader("Referer");
+    	model.addAttribute("previousUrl",previousUrl);
         System.out.println("로그인 페이지 이동: " + request.getRequestURI());
         return "login/loginPage";
     }
@@ -43,7 +48,8 @@ public class LoginController {
         String email = request.getParameter("email");
         String password = request.getParameter("password");
         String remember = request.getParameter("remember");
-
+        String previousUrl = request.getParameter("previousUrl");
+        
         System.out.println("email : " + email);
         System.out.println("password : " + password);
 
@@ -64,6 +70,14 @@ public class LoginController {
 
                 GradeDto userGrade = memberMapper.getGradeByMemberCode(member.getMem_code());
                 
+                if (member.getMem_type().equals("탈퇴")) {
+					System.out.println("탈퇴한 회원입니다.");
+					model.addAttribute("withdraw","탈퇴한 회원입니다. 정보를 복구하시겠습니까?");
+					model.addAttribute("member", member);
+					return "/login/loginPage";
+				}
+                
+                memberMapper.updatelogdate(member.getMem_code()); // 로그인시 로그데이트 현재로 업데이트
                 memberMapper.deleteWindowPro(member.getMem_code()); //유저로그인시 둘러본상품목록 전체삭제
                 
                 session.setAttribute("loginUser", member); // 세션에 사용자 정보 저장
@@ -86,7 +100,7 @@ public class LoginController {
                     System.out.println("쿠키 삭제됨 / 삭제된 이메일 : " + email);
                 }
 
-                return "redirect:/";  // 로그인 성공 시 메인 페이지로 이동
+                return "redirect:"+previousUrl;  // 로그인 성공 시 메인 페이지로 이동
             } else {
                 System.out.println("비밀번호가 일치하지 않습니다.");
                 model.addAttribute("error", "이메일 또는 비밀번호가 잘못되었습니다.");
@@ -138,6 +152,28 @@ public class LoginController {
            model.addAttribute("error", "비밀번호 변경 중 오류가 발생했습니다.");
            return "login/changePw";  // 에러 발생 시 비밀번호 변경 페이지로 다시 이동
        }
+   }
+   
+   @GetMapping("/withdraw")
+   public String withdraw(HttpSession session, HttpServletResponse response, RedirectAttributes re) {
+	   // 쿠키 삭제
+	    Cookie emailCookie = new Cookie("email", null);
+	    emailCookie.setMaxAge(0); // 쿠키 유효 기간을 0으로 설정하여 무효화
+	    emailCookie.setPath("/"); // 설정된 경로와 일치시켜야 삭제됨
+	    response.addCookie(emailCookie);
+	    MemberLoginDto member = (MemberLoginDto) session.getAttribute("loginUser");
+	    memberMapper.withdraw(member.getMem_code());
+	    re.addFlashAttribute("message","회원 탈퇴가 완료되었습니다.");
+	    session.invalidate();
+	    return "redirect:/";
+   }
+   
+   @PostMapping("/restoration")
+   public String restoration(@RequestParam("code") String memCode, RedirectAttributes re) {
+	   System.out.println(memCode);
+	   memberMapper.deleteRestoration(memCode);
+	   re.addFlashAttribute("message","복구가 완료되었습니다. 다시 로그인 해주세요");
+	   return "redirect:/";
    }
    
 }
